Send velocity as int and use pitch as uint8_t for transferring esp-now messages

+++ guitar-midi-common/midi_common.h
@@ -23,7 +23,7 @@ enum Mode {
 struct ControllerMessage {
   DeviceType device;
   Mode mode;
-  int pitch;
+  uint8_t pitch;
   int velocity;
 };
